Change build cherry-pick merge strategy to subtree.
This fixes the problem where missing files in the new build would
not be removed from the branch. For instance, if you removed a
module in the latest build, it would not get removed from the
builds repo after cherry-picking the latest build because there
was not conflict (silent error). Subtree merge strategy adds an
extra step to first adjust the new build so that it can be
compared to the existing builds apples-to-apples.