Provide build cherry-pick condition for the first build.
Subtree strategy only works well if a full site build is already
in place, otherwise it chokes with strange behavior. The subtree
strategy was used to handle the need to remove missing files no
longer present in the new build. Using the recursive strategy with
the theirs option for the first build ensures a solid baseline is
setup that will accommadate subtree strategy in subsequent builds.